implemented change of state

+++ src/main/java/co/edu/unal/software_engineering/labs/model/Balance.java
@@ -58,6 +58,6 @@ private Association association;
     }
 
     public double getPercentage(){
-        return this.balance_payed/this.total_cost;
+        return ((double)this.balance_payed)/this.total_cost;
     }
 }
